Removing the StyleRuntime and integrating the style functionalities to LicitRuntime
